requirements.txt
flask
flask-jwt-extended
flask-sqlalchemy
flask_marshmallow
export
marshmallow_sqlalchemy
